refactor: 파일 구조 변경 및 데모과정 수정
